🚚 refactor: refactor with hooks
