+++ src/features/Conversation/Actions/Error.tsx
@@ -2,7 +2,7 @@ import { ActionIconGroup } from '@lobehub/ui';
 import { ActionsBarProps } from '@lobehub/ui/es/ChatList/ActionsBar';
 import { memo } from 'react';
 
-import { useChatListActionsBar } from '../components/ChatList/useChatListActionsBar';
+import { useChatListActionsBar } from '../hooks/useChatListActionsBar';
 
 export const ErrorActionsBar = memo<ActionsBarProps>(({ text, onActionClick }) => {
   const { regenerate, del } = useChatListActionsBar(text);
