+++ src/features/Conversation/Actions/Fallback.tsx
@@ -2,7 +2,7 @@ import { ActionIconGroup } from '@lobehub/ui';
 import { memo } from 'react';
 
 import { RenderAction } from '../components/ChatList';
-import { useChatListActionsBar } from '../components/ChatList/useChatListActionsBar';
+import { useChatListActionsBar } from '../hooks/useChatListActionsBar';
 
 export const DefaultActionsBar: RenderAction = memo(({ text, onActionClick }) => {
   const { del } = useChatListActionsBar(text);
