+++ src/features/Conversation/Actions/Function.tsx
@@ -2,7 +2,7 @@ import { ActionIconGroup } from '@lobehub/ui';
 import { memo } from 'react';
 
 import { RenderAction } from '../components/ChatList';
-import { useChatListActionsBar } from '../components/ChatList/useChatListActionsBar';
+import { useChatListActionsBar } from '../hooks/useChatListActionsBar';
 
 export const FunctionActionsBar: RenderAction = memo(({ text, onActionClick }) => {
   const { regenerate, divider, del } = useChatListActionsBar(text);
