+++ src/features/Conversation/Actions/User.tsx
@@ -2,7 +2,7 @@ import { ActionIconGroup } from '@lobehub/ui';
 import { memo } from 'react';
 
 import { RenderAction } from '../components/ChatList';
-import { useChatListActionsBar } from '../components/ChatList/useChatListActionsBar';
+import { useChatListActionsBar } from '../hooks/useChatListActionsBar';
 import { useCustomActions } from './customAction';
 
 export const UserActionsBar: RenderAction = memo(({ text, onActionClick }) => {
